backend/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
PyMongo[srv]==4.13.2
python-dotenv==1.1.1
bcrypt==4.3.0
Flask-JWT-Extended==4.7.1
gunicorn==21.2.0 